Add a last modified date to single layout
Instead of manually tracking when the now page was modified, this change
will make it use the date of the last commit that touched the file.

Additionally, this commit adds the last modified date to all single
pages (except the homepage) so that book notes and articles also show
the last modified date.