Consistent categories between books and articles
Originally, I thought I was going to write articles only for programming
but I remembered that I already have some articles written on Medium
that I want to migrate.

These old articles are not programming specific so I decided to keep a
consistent category structure between books and programming because, at
the end of the day, that's what's interesting for me.